requirements.txt
# YouTube downloading and processing
yt-dlp>=2024.12.0

# RSS/Podcast feed generation
feedgen>=1.0.0

# Audio metadata extraction
mutagen>=1.47.0

# GitHub API interaction
PyGithub>=2.1.1

# Configuration file parsing
PyYAML>=6.0.1

# HTTP requests (fallback for GitHub API)
requests>=2.31.0

# Date/time utilities
python-dateutil>=2.8.2
